#84 refactor: product detail 이사합니당
