+++ src/main/java/com/webeye/backend/imageanalysis/infrastructure/OpenAiClient.java
@@ -4,12 +4,11 @@ import com.webeye.backend.allergy.dto.response.AllergyAiResponse;
 import com.webeye.backend.cosmetic.dto.response.CosmeticResponse;
 import com.webeye.backend.imageanalysis.dto.request.ImageAnalysisRequest;
 import com.webeye.backend.imageanalysis.dto.response.ImageAnalysisResponse;
-import com.webeye.backend.product.dto.request.ProductAnalysisRequest;
 import com.webeye.backend.product.dto.response.DetailExplanationResponse;
 import com.webeye.backend.global.error.BusinessException;
 import com.webeye.backend.healthfood.dto.HealthFoodAiResponse;
 import com.webeye.backend.imageanalysis.dto.request.ImageAnalysisPrompt;
-import com.webeye.backend.product.domain.type.OutlineType;
+import com.webeye.backend.productdetail.domain.type.OutlineType;
 import com.webeye.backend.product.dto.request.FoodProductAnalysisRequest;
 import com.webeye.backend.nutrition.dto.response.NutritionAiResponse;
 import com.webeye.backend.rawmaterial.dto.response.RawMaterialAiResponse;
