+++ src/main/java/com/webeye/backend/product/presentation/swagger/ProductSwagger.java
@@ -2,7 +2,7 @@ package com.webeye.backend.product.presentation.swagger;
 
 import com.webeye.backend.product.dto.response.DetailExplanationResponse;
 import com.webeye.backend.global.dto.response.SuccessResponse;
-import com.webeye.backend.product.domain.type.OutlineType;
+import com.webeye.backend.productdetail.domain.type.OutlineType;
 import com.webeye.backend.product.dto.request.FoodProductAnalysisRequest;
 import com.webeye.backend.product.dto.request.ProductDetailAnalysisRequest;
 import com.webeye.backend.product.dto.response.ProductResponse;
